requirements.txt
anyio==3.6.1
APScheduler==3.6.3
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
h11==0.13.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
mysql-connector-python==8.0.30
outcome==1.2.0
protobuf==3.20.1
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tornado==6.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
